After converting from svn, normalize the eol
Git and SVN have different ways to handle end of lines,
the conversion between CRLF and LF.
Typically Git under Windows has core.autocrlf = true,
but Git under Cygwin or Linux/Unix has not.

A common view on text files vs binary files is needed,
regardless of the local core.autocrlf.
This can be done using a .gitattribute file, which is checked in and
travels to the different clones using git clone.

After the conversion fron svn, convert all files to have LF in the repo,
(binary files are not touched), and commit the changes in a separate commit.

More details can be seen here:
https://www.kernel.org/pub/software/scm/git/docs/gitattributes.html

Note that *.cmd can be used for different things:
as dosomething.cmd file in Windoes (where CRLF is needed)
or as st.cmd under EPICS (where probaly LF is prefered)
Mark the cmd files as not text and switch of the conversion